ACT-431:- Clean up Indicators app admin.py file
- Use decorators to register admin models
- Remove all admin models from models.py file to admin.py file
